Add a private repo whitelist
We can now allow the publishing of private repositories. Currently this
is static and in the code itself, if we need to move this into config or
a database later then we can do.

I'm going to check the existing database for any private repositories
next, so that we don't miss any.

This resolves #38